Add pipes and methods for errors into products controller

+++ server/src/products/products.controller.ts
@@ -1,35 +1,46 @@
 import {
-  Controller, Body, Param, 
-  Get, Post, Put, Delete
+  Controller, Body, Param,
+  Delete, Get, Post, Put,
+  UsePipes, NotFoundException, BadRequestException,
 } from '@nestjs/common';
 import { CreateProductDto } from './dto/create-product.dto';
+import { ValidationPipe } from './pipes/validation.pipe';
 import { ProductsService } from './products.service';
 
 @Controller('products')
 export class ProductsController {
   constructor(private productsService: ProductsService) {}
 
+  @UsePipes(ValidationPipe)
   @Post()
-  async createProduct(@Body() dto: CreateProductDto) {
+  createProduct(@Body() dto: CreateProductDto) {
     return this.productsService.createProduct(dto);
   }
 
   @Get()
   async getProductsList() {
     const productsList = await this.productsService.getProductsList();
+    if (!productsList) throw new NotFoundException();
+
     return productsList;
   }
 
   @Get(':id')
   async getProduct(@Param('id') id: number) {
     const product = await this.productsService.getProduct(id);
+    if(!product) throw new NotFoundException();
+
     return product;
   }
 
+  @UsePipes(ValidationPipe)
   @Put(':id')
   async updateProduct(
     @Param('id') id: number,
     @Body() dto: CreateProductDto) {
+    const product = await this.productsService.getProduct(id);
+    if(!product) throw new BadRequestException();
+
     this.productsService.updateProduct(id, dto);
 
     return `Product ${id} was updated`;
@@ -38,6 +49,9 @@ export class ProductsController {
   @Delete(':id')
   async removeProduct(
     @Param('id') id: number) {
+    const product = await this.productsService.getProduct(id);
+    if(!product) throw new BadRequestException();
+
     this.productsService.removeProduct(id);
 
     return `Product ${id} was removed`;
